pyproject.toml
[project]
name = "django-starter-template"
version = "0.1.0"
description = "A simple way to start your Django project with a solid foundation."
readme = "README.md"
authors = [{ name = "difegam", email = "difegam3@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-allauth[socialaccount]>=65.13.1",
    "django-environ>=0.12.0",
]


[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "django-stubs-ext>=5.2.7",
    "django-stubs[compatible-mypy]>=5.2.7",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.0",
]

[tool.djlint]
profile = "django"

[tool.codespell]
skip = "migrations"
